+++ src/targets/adminGetUser.test.ts
@@ -47,6 +47,6 @@ describe("AdminGetUser target", () => {
         Username: existingUser.Username,
         UserPoolId: "test",
       })
-    ).rejects.toEqual(new UserNotFoundError("User does not exist"));
+    ).rejects.toEqual(new UserNotFoundError("User does not exist."));
   });
 });